+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Output a notice if exec'ing a SUID or static-pie executable
 - Preload the iamroot library not using `IAMROOT_LIB` if `LD_PRELOAD` is unset
   and not in a chroot'ed environment
+- Strip leading `ld_library_path` from the content of `LD_LIBRARY_PATH`
 
 ### Fixed